+++ unique_toolkit/unique_toolkit/app/__init__.py
@@ -23,3 +23,5 @@ from .schemas import (
 from .verification import (
     verify_signature_and_construct_event as verify_signature_and_construct_event,
 )
+
+DOMAIN_NAME = "app"