+++ unique_toolkit/unique_toolkit/content/__init__.py
@@ -35,3 +35,5 @@ from .utils import (
 from .utils import (
     sort_content_chunks as sort_content_chunks,
 )
+
+DOMAIN_NAME = "content"