+++ unique_toolkit/unique_toolkit/embedding/__init__.py
@@ -3,3 +3,5 @@ from .service import EmbeddingService as EmbeddingService
 from .utils import (
     calculate_cosine_similarity as calculate_cosine_similarity,
 )
+
+DOMAIN_NAME = "embedding"