+++ unique_toolkit/CHANGELOG.md
@@ -5,7 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2024-12-17
+- remove the `Event` dependency from `LanguageModelService`
+- move SDK calls to `unique_toolkit.language_model.functions`
+- ...
 
+## [0.5.43] - 2024-12-13
+- Add `Prompt` class to handle templated prompts that can be formatted into LanguageModelSystemMessage and LanguageModelUserMessage.
+
+## [0.5.42] - 2024-12-11
+- Update `LanguageModelTokenLimits` with fix avoiding floats for token
+
+## [0.5.41] - 2024-12-11
+- Update `LanguageModelTokenLimits` includes a fraction_input now to always have input/output token limits available.
+
+## [0.5.40] - 2024-12-11
+- Add `other_options` to `LanguageModelService.complete`, `LanguageModelService.complete_async`, `LanguageModelService.stream_complete` and `LanguageModelService.stream_complete_async`
 
 ## [0.5.39] - 2024-12-09
 - Add `contentIds` to `Search.create` and `Search.create_async`